Moved activityPub methods into their own module and added stubs for future methods
